Backend/requirements.txt
fastapi
uvicorn
pandas
numpy
faiss-cpu
scikit-learn
datasets
cachetools
pyarrow
python-dotenv
google-genai